+++ src/index.ts
@@ -4,13 +4,14 @@
  * @git             https://github.com/devsdaddy/quarkdash
  * @version         1.1.0
  * @author          Elijah Rastorguev
- * @build           1002
+ * @build           1003
  * @website         https://dev.to/devsdaddy
- * @updated         13.04.2026
+ * @updated         14.04.2026
  */
 /* Export Types and Utils */
 export * from "./core/types";
 export * from "./core/utils";
+export * from "./core/wasm_loader";
 
 /* Export KDF and MAC */
 export * from "./core/kdf";